==> code/1_Extracting Data & Installing Packages.ipynb <==
# Downloading Data

from os.path import getsize
from urllib.request import urlretrieve

Download Data of High Volume For Hire Vehicles from 09/2020 to 12/2020 

# Codes are based on tutorial materials

output_dir = "/Volumes/E/data"
fname_template = "fhvhv_tripdata_2020"

for m in range(9,13):
    month = str(m).zfill(2)
    out = f'{fname_template}-{month}.csv'
    url = f"https://nyc-tlc.s3.amazonaws.com/trip+data/{out}"
    urlretrieve(url, f"{output_dir}/{out}")

    print(f"Done downloading {out} to {output_dir} with size {getsize(f'{output_dir}/{out}') / 1073741824:.2f}GB")
   

Download COVID-19 data

# Data recording number of cases by day since the outbreak of covid-19

output_dir = "/Volumes/E/data"
fname = "cases-by-day.csv"

url = f"https://raw.githubusercontent.com/nychealth/coronavirus-data/master/trends/{fname}"
urlretrieve(url, f"{output_dir}/{fname}")

print(f"Done downloading {fname} to {output_dir} with size {getsize(f'{output_dir}/{fname}') / 1024:.2f}KB")


# Data serialisation

import pandas as pd
import warnings
warnings.filterwarnings("ignore")


fhvhv_0920 = pd.read_csv('/Volumes/E/data/fhvhv_tripdata_2020-09.csv')
fhvhv_0920.to_feather('/Volumes/E/data/feather/fhvhv_tripdata_2020-09.feather')

fhvhv_1020 = pd.read_csv('/Volumes/E/data/fhvhv_tripdata_2020-10.csv')
fhvhv_1020.to_feather('/Volumes/E/data/feather/fhvhv_tripdata_2020-10.feather')

fhvhv_1120 = pd.read_csv('/Volumes/E/data/fhvhv_tripdata_2020-11.csv')
fhvhv_1120.to_feather('/Volumes/E/data/feather/fhvhv_tripdata_2020-11.feather')

fhvhv_1220 = pd.read_csv('/Volumes/E/data/fhvhv_tripdata_2020-12.csv')
fhvhv_1220.to_feather('/Volumes/E/data/feather/fhvhv_tripdata_2020-12.feather')

# Installing Packages

!pip install numpy

!pip install folium

!pip install seaborn

!pip install geopandas

!pip install statsmodels
