==> code/2_Preprocessing & Exploratory Data Analysis.ipynb <==
# Reading and merging data

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings("ignore")

# Data from 09/2020 to 12/2020 for high volume fhv

fhvhv_0920_feather = pd.read_feather('/Volumes/E/data/feather/fhvhv_tripdata_2020-09.feather')
fhvhv_1020_feather = pd.read_feather('/Volumes/E/data/feather/fhvhv_tripdata_2020-10.feather')
fhvhv_1120_feather = pd.read_feather('/Volumes/E/data/feather/fhvhv_tripdata_2020-11.feather')
fhvhv_1220_feather = pd.read_feather('/Volumes/E/data/feather/fhvhv_tripdata_2020-12.feather')


# Merge the data from Sep to Dec

frames = [fhvhv_0920_feather, fhvhv_1020_feather, fhvhv_1120_feather, fhvhv_1220_feather]
fhvhv_2020_df = pd.concat(frames, keys=["Sep", "Oct", "Nov", "Dec"])
fhvhv_2020_df.head()

# Data that records covid cases by day in nyc

covid_df = pd.read_csv('/Volumes/E/data/cases-by-day.csv')
covid_df.head()

# Data cleaning and preprocessing

High Volume FHV data

fhvhv_2020_df.dtypes

fhvhv_2020_df.drop(columns=['dispatching_base_num'], inplace=True)

# Impute missing values for share rides 
fhvhv_2020_df['SR_Flag'].fillna(0, inplace=True)
fhvhv_2020_df.dropna(0, inplace = True)
print("Done filling missing values")

# Convert share rides to categorical data
fhvhv_2020_df['SR_Flag'] = fhvhv_2020_df['SR_Flag'].astype('bool')

# Check if any rows have missing values, if not, proceed
fhvhv_2020_df.isnull().sum().sort_values(ascending=False)

# Add a new column which comprises the duration of each trip in minutes
# Codes are modified based on:
# https://stackoverflow.com/questions/51938140/how-to-get-the-time-duration-from-two-date-time-columns-of-pandas-dataframe
# Might take up to 10 minutes to load...

covert_to_min = 60

fhvhv_2020_df['pickup_datetime'] = pd.to_datetime(fhvhv_2020_df['pickup_datetime'])
fhvhv_2020_df['dropoff_datetime'] = pd.to_datetime(fhvhv_2020_df['dropoff_datetime'])
fhvhv_2020_df['trip_duration_min'] = (fhvhv_2020_df['dropoff_datetime'] - fhvhv_2020_df['pickup_datetime']).apply(
    lambda x: x.seconds/(covert_to_min))
fhvhv_2020_df['trip_duration_min'] = fhvhv_2020_df['trip_duration_min'].round(2)

fhvhv_2020_df.head()

# Remove outliers: minimum and maximum values of trip duration

outlier_min = fhvhv_2020_df['trip_duration_min'].min()
print('Minimum trip duration: ',outlier_min)
outlier_max = fhvhv_2020_df['trip_duration_min'].max()
print('Maximum trip duration: ',outlier_max)

fhvhv_2020_df = fhvhv_2020_df[(fhvhv_2020_df.trip_duration_min != outlier_min)
                              & (fhvhv_2020_df.trip_duration_min != outlier_max)]    
print('Outliers removed!')

# Generate a new column pick_day indicating the day of pickup

fhvhv_2020_df['pickup_day'] = pd.DatetimeIndex(fhvhv_2020_df['pickup_datetime']).weekday


COVID-19 data

# Check data type
covid_df.dtypes

# Drop this column as it is irrelevant to analysis
covid_df.drop(['INCOMPLETE'], axis=1, inplace=True)

# Only keep the data within the time period of interest

start_date = '2020-09-01'
end_date = '2020-12-31'

covid_df['date_of_interest'] = pd.to_datetime(covid_df['date_of_interest'])
time_period = (covid_df['date_of_interest'] >= start_date) & (covid_df['date_of_interest'] <= end_date)
covid_df = covid_df.loc[time_period].reset_index(drop=True)

# Descriptive Statistics

High Volume FHV data

# Check dimension of the data
fhvhv_2020_df.shape

pd.options.display.float_format = '{:,.2f}'.format
fhvhv_2020_df.describe()

# Count the number of trips according to HVFHS business
# HV0002: Juno
# HV0003: Uber
# HV0004: Via
# HV0005: Lyft

print("Number of trips according to HVFHS business from 09/2020 to 12/2020:")
fhvhv_2020_df['hvfhs_license_num'].value_counts()

COVID-19 data

covid_df.describe()

# Data aggregation

# Prepare to aggregate fhvhv taxi data and covid data
# Takes about 5 min to load

# Keep attributes we are interested in
fhvhv_by_day_df = fhvhv_2020_df[['pickup_datetime','SR_Flag','trip_duration_min']]

# To match the datetime template in covid data
fhvhv_by_day_df['pickup_date'] = fhvhv_2020_df['pickup_datetime'].dt.strftime('%Y-%m-%d')

fhvhv_by_day_df.head()

# Aggregating all attributes by date
# Codes are modified based on:
# https://stackoverflow.com/questions/45752601/how-to-do-a-conditional-count-after-groupby-on-a-pandas-dataframe


reduced_df = fhvhv_by_day_df.groupby('pickup_date')['SR_Flag'].apply(
    lambda x: (x==True).sum()).reset_index(name='num_of_SR')
reduced_df['trip_count'] = fhvhv_by_day_df.groupby('pickup_date')['pickup_date'].agg('count').reset_index(
    name='trip').trip
reduced_df['avg_trip_duration'] = fhvhv_by_day_df.groupby('pickup_date')['trip_duration_min'].agg('mean').reset_index(
    name='duration').duration
reduced_df.head()

# Combine the fhvhv data with covid data

fhvhv_covid_df = pd.merge(reduced_df,covid_df,left_index=True,right_index=True)
fhvhv_covid_df = fhvhv_covid_df.drop(columns=['date_of_interest'])
fhvhv_covid_df.describe()

# Saving preprocessed data

fhvhv_2020_df = fhvhv_2020_df.reset_index(drop=True)
fhvhv_2020_df.to_feather(r'/Volumes/E/data/preprocessed_data/fhvhv_2020.feather')

fhvhv_covid_df.to_csv(r'/Volumes/E/data/preprocessed_data/fhvhv_covid.csv',index=False)
