==> code/4_Statistical Modelling.ipynb <==
# Reading preprocessed data

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings("ignore")

fhvhv_covid_df = pd.read_csv('/Volumes/E/data/preprocessed_data/fhvhv_covid.csv')
fhvhv_covid_df

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import Ridge
from sklearn.metrics import mean_squared_error,mean_absolute_error 
from sklearn.metrics import r2_score

from numpy import absolute
from numpy import mean
from numpy import std

from statsmodels.formula.api import ols,glm
from statsmodels.api import families

# train set: data from Sep to Nov
train_df = fhvhv_covid_df.loc[(fhvhv_covid_df['pickup_date'] >= '2020-09-01') & 
                          (fhvhv_covid_df['pickup_date'] <= '2020-11-30')]
train_df.describe()

# test_set: Dec data 
test_df = fhvhv_covid_df.loc[(fhvhv_covid_df['pickup_date'] >= '2020-12-01') & 
                          (fhvhv_covid_df['pickup_date'] <= '2020-12-31')]
test_df.describe()

# Fitting ridge regression model

X = train_df.drop(['pickup_date','trip_count'],axis=1).values
y = train_df['trip_count'].values
y = y.reshape(-1, 1)

# Standardize data
scaler = StandardScaler()
scaler.fit_transform(X)
X = scaler.transform(X)
scaler.fit(y)
y = scaler.transform(y)


# Split train and test set
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.33, random_state=1)

rr = Ridge(alpha=10)
rr.fit(X_train, y_train)
pred_train = rr.predict(X_train)

# Compute MAE, MSE, R2 to evalute the performance of the model
print("MAE on train data: ", mean_absolute_error(y_train, pred_train))
print("MSE on train data: ", mean_squared_error(y_train, pred_train))
print("R2 score on train data: ", r2_score(y_train, pred_train))

pred_test = rr.predict(X_test)
print("\nMAE on test data: ", mean_absolute_error(y_test, pred_test))
print("MSE on test data: ", mean_squared_error(y_test, pred_test))
print("R2 score on test data: ", r2_score(y_test, pred_test))

# Hyperparameter tuning

# Grid search

from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import RepeatedKFold

# Define model evaluation method
cv = RepeatedKFold(n_splits=10, n_repeats=3, random_state=1)
# Define grid
grid = {'alpha':[0.01, 0.1, 1, 10]}

search = GridSearchCV(rr, grid, scoring='neg_mean_absolute_error', cv=cv, n_jobs=-1)

results = search.fit(X, y)

print('MAE: %.3f' % results.best_score_)
print('Config: %s' % results.best_params_)

# Predicting future data

X = test_df.drop(['pickup_date','trip_count'],axis=1).values
y = test_df['trip_count'].values
y = y.reshape(-1, 1)

# Standardize data
scaler.fit(X)
X = scaler.transform(X)
scaler.fit(y)
y = scaler.transform(y)

pred_y = rr.predict(X)

# Compute MAE, MSE, R2 to evalute the performance of the model
print("MAE on future data: ", mean_absolute_error(y, pred_y))
print("MSE on future data: ", mean_squared_error(y, pred_y))
print("R2 score on future data: ", r2_score(y, pred_y))
